:wrench: Improve usage of task lists for project issue and MR templates
Leveraged the innaplicable tasks feature of GitLab Flavoured Markdown to
require contributors to better specify binary options. This was always
the intended use for the tasks lists, but a rendering bug made us opt
for a less robust approach.

Now that gitlab.org/gitlab#578684 has been opened, we have decided to
implement the original vision for task lists, regardless if the issue
will be picked by the GitLab team or not.

Changelog: policies