:pencil: Add template badges for quick navigation
Added badges for common queries in the issue tracker in the template
`CONTRIBUTING.md` for visual cues to important filters.

Additionally, replaced generic external links for badges in `README.md`
with links pointing to specific sections of the guidelines. Removed the
unnecessary sections on Galactipy features and suggestions.

Closes #178, #263, #323, #347, #348, #349, #350, #351, #359, #385

Changelog: template-config