:art: Improve Docker login task for GitLab
Fixed a bug mandating the usage of the `password` argument for Docker
Hub logins, which is preferably done via one-time device login.

Also took the opportunity to improve security on the GitLab task,
structuring it in a way that is can only be run through a pipe command
and added a PAT regex validation.

Changelog: bug