:pencil: Update project `README.md` with quality fixes
Converted the remaining sections of the project `README.md` to SemBr and
made minor fixes to Markdown rendering and descriptions.

The most significant changes were made to the "What's Next" section,
leveraging GitLab Flavoured Markdown and revising the list of
recommended libraries and readings now that the project has finally
reached the stable release.

Closes #154, #317, #391, #394, #428

Changelog: documentation