:art: Refine terms used for licences
The template `README.md` used the bare `licence` variable, which is the
SPDX identifier and not human-friendly. Revamped the `cookiecutter.json`
file to deal with licence naming directly in the configuration file and
avoiding the usage of per-file mappings and variables.

Additionally, modified `CITATION.cff` to use `license-url` instead for
non-OSS licences.

Changelog: hooks