:sparkles: Implement trusted publishing for PyPI
The SLSA CI/CD component only works by generating single-file
attestations. Given that versions are dynamic and there is not a
straightforward way to remove the "v" prefix of the version tag without
setting up an additional job, the solution has been overlooked in favour
of following the PyPI recommended settings for publishing packages in a
trusted environment.

Closes #486, #491

Changelog: feature