:sparkles: Implement uploads to the GitLab Package Registry
Added a job to manually publish any tag (pre-release or release) to the
repository's GitLab Package Registry, using the PyPI version to name the
package.

An `after_script` section is included to retrieve and store the package
ID, in order to use it for the release links later.

Closes #443

Changelog: feature